+++ examples/src/js/FilterExample.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import CheckboxTree from 'react-checkbox-tree';
 
-import { fileSystem as nodes } from './common';
+import { fileSystem as nodes } from './common.js';
 
 class FilterExample extends Component {
     state = {