+++ src/js/lang/default.js
@@ -1,0 +1,6 @@
+export default {
+    collapseAll: 'Collapse all',
+    collapseNode: 'Collapse node',
+    expandAll: 'Expand all',
+    expandNode: 'Expand node',
+};