+++ src/js/shapes/languageShape.js
@@ -1,7 +1,8 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
-    collapseAll: PropTypes.string.isRequired,
-    expandAll: PropTypes.string.isRequired,
-    toggle: PropTypes.string.isRequired,
+    collapseAll: PropTypes.string,
+    collapseNode: PropTypes.string,
+    expandAll: PropTypes.string,
+    expandNode: PropTypes.string,
 });