+++ test/index.js
@@ -1,0 +1,5 @@
+import './components/Button.jsx';
+import './components/CheckboxTree.jsx';
+import './components/NativeCheckbox.jsx';
+import './components/TreeNode.jsx';
+import './utils.js';